Fix all warnings in Xcode 7.1
Add conditional compilation around save code that uses
NSConfinementConcurrencyType, since that is deprecated on iOS 9 but
still needed if we're targetting an earlier OS. This silences the
warnings without sacrificing compatibility.

Also, add an override for -init since it is now marked as a designated
initializer on NSObject.

And fix a spurious Framework search path in the Tests target.